requirements.txt
lightllm[proxy]
pyautogen